requirements.txt
fastapi
google-generativeai
uvicorn
python-multipart
python-docx
pandas
openpyxl
beautifulsoup4
markdown
SpeechRecognition
moviepy
Pillow
python-dotenv
sentence-transformers
faiss-cpu
pythainlp
pydub
einops
pypdf